seperate the wasm instructions from runtime code add support for .so dylib files implement a missing wasi_intruction (prestat)
