+++ runtime/libc/cortex_m_backing.c
@@ -4,6 +4,17 @@ int printf_(const char* format, ...);
 
 volatile int CORTEX_M_ARG_C = 0;
 
+u32 allocate_n_bytes(u32 n) {
+    awsm_assert(memory_size > 0);
+    u32 res = runtime_heap_base;
+    runtime_heap_base += n;
+    while (memory_size < runtime_heap_base) {
+        expand_memory();
+    }
+    printf("rhb %d\n", runtime_heap_base);
+    return res;
+}
+
 int runtime_main(int argc, char** argv) {
     runtime_init();
